Use ExternalLinks and externalUrls for PubMed in LiteratureCitation. Add EuropePMC to externalUrls. Extract getCitationItemId into own util file.
